Fixed balance calcs

+++ utils.h
@@ -7,7 +7,7 @@ struct Account{
     int id;
     char name[20];
     char password[20];
-    float balance;
+    long long int balance;
 };
 
 void loggedIn(struct Account* currentAccount, struct Account* lastLoggedInAccount, bool isLoggedIn);
